% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa-methods.R
\name{RV}
\alias{RV}
\title{RV}
\usage{
RV(tbl1, tbl2)
}
\arguments{
\item{tbl1}{a matrix or data frame}

\item{tbl2}{a matrix or data frame}
}
\description{
Compute the similarity between two tables using the Rv coefficient
}

