% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa-methods.R
\name{RVtbl}
\alias{RVtbl}
\title{RVtbl}
\usage{
RVtbl(data, sets)
}
\arguments{
\item{data}{data set (matrix or data frame)}

\item{sets}{list of vectors indicating the sets of variables}
}
\description{
Compute the similarity matrix between multiple sets of tables
}

